prefer common approach to Promise construction
now Promises are constructed, as in most implementations, by passing the initializer a block which is called with the fulfill function as an argument. Promise no longer supports being fulfilled from anywhere via the <- operator. this should encourage better design in most cases and when the old behavior is needed it's easy enough to break the fulfill function out of the initial block.