rework PromiseSequence API & rename some things
as with Promise previously, abstracted out the parts that don't need to know about Task and moved them up a level. that includes RecurringPromise and PromiseSequence, which is now called Feed. PromiseSequenceGenerator is replaced with Subscription, which differs in usage in that it is a struct instead of a class; hence, Subscriptions are passed by value, and saving or passing a Subscription value can be used to save a reference to a particular point in a Feed. the Feed itself is an object and holding a reference to it can always be used to obtain a Subscription starting at the current head.